deps(backend): adopt sqlite package
Instead of treating our SQLite bindings as a fork of crawshaw.io/sqlite we now adopt the package directly into our code base.
The upstream is abandoned, and we've made modifications to it beyond the point of converging the forks.

We'll be making even more changes in the face of new features introduced in Go 1.23, so it make sense to adopt the package now and start ownging it entirely.