refactor: allow skipping the daemon spawning in the desktop app
When we want quick feedback loop iterating on the backend, we don't want to wait
for the desktop app to rebuild over and over.

This commit allows running the app with `SEED_NO_DAEMON_SPAWN=1` env variable,
and run the backend in a separate process with `go run` or similar.