fix(backend): implement blob stashing + Refs fixes
We used to be very tolerant to errors when indexing blobs other than Changes,
but those errors were never tracked anywhere. Now we have a separate table
where we can stash those blobs that we started indexing, but couldnt't finish for whatever reason.

We also now trying to reindex those stashed blobs when we believe those errors might be resolved.

Mainly this means that Refs are being reindexed when we receive Capabilities that cover previously denied Ref,
or when we receive a Change that was a previously missing Head for a Ref.