backend(fix): fix the encoding of comments
We screwed up the encoding of blocks and annotations inside comments.
Instead of having untyped properties inline they were inside the "attributes" field,
and the ID was spelled as "iD" instead of "id" due to some gotchas in the encoding libraries we use.

This commit fixed the encoding for new comments, without breaking the old comments.