fix(backend): avoid redundant block replacements in the encoded changes
Some time ago we've introduced a regression, which caused changes to encode ReplaceBlock operations even in situations where block is not touched at all.
This is partially caused by our frontend not deduplicate the redundant ops, to simplify the diffing two document states. But the backend used to handle it gracefully,
until something happened that broke it. This commit fixes it.