fix(backend): fix block reordering problem
Sometimes blocks were reordering when documents are published which is very frustrating for users.
There was an error in the batching of CRDT ops which caused some operations to be batched into one sequence.

It is partially caused by our frontend not tracking user changes precisely, and instead diffing documents to derive the operations,
which causes some redundant operations that the backend is not interpreting correctly.

Fixes https://linear.app/seedhypermedia/issue/SHM-762/blocks-order-break-when-leftsibling-is-a-new-block.