fix(backend): fix old document generations show up in the list after deletion
Sometimes deleting the latest generations could resurface old generations in the list,
while GetDocument for them would be failing. This was causing annoying situations where a document would appear in the list,
but when you click on it, it would get stuck in the discovery phase which was very confusing.

Closes: https://linear.app/seedhypermedia/issue/SHM-962/cannot-delete-document-from-site.