frontend(device-linking): fix device linking
Sometimes our libp2p dials would get stuck, because we would connect to the relay server, but wouldn't connect to the actual target peer.
So, for some reason js-libp2p just hangs in this case.

This commit adds explicit timeouts for dials, and provides a bit more feedback to the user for what's going on.

Unfortunately our DX with Remix at the moment is really bad, and requires full rebuilds, so I added a debug web page which uses the same device linking logic but it's much easier to run in development with fast feedback loop using Bun. This should make it easier for us to debug the device linking process locally.