feat(backend): expose comment IDs for linking
This commit brings us closer to our [Vision for Addressability](https://linear.app/seedhypermedia/document/vision-for-addressability-32e5fedddff8)
by exposing record IDs on comments, in addition to version, which builds up to being able to address comments with URLs.

In the API layer the ID of a comment is a Record ID string, which the format of `{authorityPublicKey}/{tsid}`.
Despite the annoying fact that it's a yet another ID format, I didn't want to make comment IDs being `hm://` URLs for various reasons:

1. I didn't want to bring confusion whether comment IDs could support URL parameters.
2. Because we don't use URLs anywhere in the document APIs, I didn't want to bring more inconsistencies just for comments.

So, the rationale is that constructing URLs for Documents, Contacts, and Comments needs to happen on the client, and the API surface doesn't undersating anything about `hm://` URLs.

At some point I want to refactor our API entirely, to make it more consistent, and more aware of the hypermedia nature of our data, and being able to resolve URLs for any data.