feat(backend/sqlite): allow nested transaction by falling back to using a savepoint
Somewhere in our code we have a problem with a nested SQLite transaction.
This strictness doesn't really help much in our case,
and sometimes it's very flaky and even platform-dependent — really weird.
It feels pretty safe to just fall back to using a SAVEPOINT when BEGIN IMMEDIATE fails due to a nested transaction,
which is what this commit implements.