fix(backend): fix missing parent for moves of children of detached blocks
We had a bug when a detached block was created along with its children in the same change.
It's because when we clean up and prepare the change we put move operations before replace operations,
which resulted in moves refering to parents that were never mentioned before.
We fixed this by just allowing this to happen.

Fixes https://linear.app/seedhypermedia/issue/SHM-1252/cannot-move-blocks-under-detached-block.