fix(backend): fix comment causality issues again
We believed that we've fixed all the problems of comments being indexed out of order,
but turns out there were more edge cases. Particularly when there're multiple levels of replies involved.

Part of the issue seems to be the fact that we were doing nested SQLite queries and select and update the same table,
which seems like is known to cause some weird behaviors. We fixed this by buffering the results from one query before running the other ones,
instead of doing N+1 queries right in the same loop.