Fix the migrations
There was a typo in table name.

The first line was `DROP TABLE IF EXISTS fts5`, but there's no table with such name.
Next line was `CREATE VIRTUAL TABLE IF NOT EXISTS fts` which didn't have any effect because the old table with this name kept existing.

That's why the migration tests were failing.