fix(backend): fix missing account aliases during discovery
We used to have the problem that account aliases were not respected, because the corresponding AGENT capabilities were missing.
This commit fixes this, along with some additional problems that used to happen because of broken causality between blobs.

We now are more resilient to causality problems, and I've added more tests to verify that.

Fixes https://linear.app/seedhypermedia/issue/SHM-1347/account-aliases-are-not-syncing.