fix(daemon): FTS returns correct document versioning
Some search items were not successfully indexed since the logic in the reindex process was odd. Also the performance on searches was poor.
We implemented a light indexing loging at the time of reindexing the database and moved the "getting the latest version of a match" back at the search time.