chore(frontend): move constants from the @shm/shared barrel file
This commit removes `constants.ts` from the `@shm/shared` barrel file in an attempt to fix our Remix dev build.
Somehow the constants file makes it into the dev build of our web app, which blows up because `process` is not defined in the browser.
I understand that we wanted to move away from the barrel files anyway, so I simply changes all the usages for those constants to import from the module directly.

The idea is that this should make it easier for us to see which other files depend on those constants, to see if some of those may need to be provided via React context instead,
to avoid the whole constants file being included into the dev build.