chore(frontend/desktop): remove the gap before the accessory panel
I complained at some point about the unnecessarily huge gap between the main panel and the accessory panel in desktop,
and apparently people were not against of reducing it, making it a bit more similar to how it's on the web.

This commit removes the gap and makes the resize handle a bit thinner too, to avoid it pushing the gap for its entire width.