feat(desktop): handle deep link URLs in the search box
Sometimes it's hard to test deep links, because the dev application doesn't register the URL handler in the operating system.
This commit handles deep links via the Launcher's search box, so deep links can be tested this way.
And it's just nice to be able to handle deep links via our "resolve anything" interface, which is the search.