feat(frontend): add tooltips with full dates for feed events
Currently we show only timestamp for the current day's events, and only days for the earlier events.

Based on most modern apps, it seems like it's a good practice to have a tooltip with the full timestamps when displaying dates, just in case one might need them. I often do.