Remove unused progress notifier
This was quite confusing and is not used anywhere.

It should be responsibility of the caller to poll the progress if needed.
Pretending to have live updates, while in fact limiting them to a interval defeats the purpose of live updates.
Polling the progress stats would work just as well.